main/requirements.txt
numpy
matplotlib
scipy
bokeh
python-pptx
python-docx
colorama